+++ MovieQuiz/QuizQuestion.swift
@@ -1,0 +1,8 @@
+//
+//  QuizQuestion.swift
+//  MovieQuiz
+//
+//  Created by Алина on 30.11.2024.
+//
+
+import Foundation
