+++ MovieQuiz/QuizResultsViewModel.swift
@@ -1,0 +1,8 @@
+//
+//  QuizResultsViewModel.swift
+//  MovieQuiz
+//
+//  Created by Алина on 30.11.2024.
+//
+
+import Foundation
